Update on "[PT-D] Enable init ops for DTensor"
[ghstack-poisoned]
